+++ src/storage/mod.rs
@@ -14,6 +14,8 @@ pub trait KvInterface: std::default::Default + IntoIterator + Import + Export {
     type Value;
     type RefKey: ?Sized;
 
+    fn with_capacity(cap: usize) -> Self;
+
     fn has_key(&self, key: &Self::RefKey) -> bool;
 
     fn is_empty(&self) -> bool;
@@ -26,7 +28,7 @@ pub trait KvInterface: std::default::Default + IntoIterator + Import + Export {
     fn get(&self, key: &Self::RefKey) -> Option<&Self::Value>;
     fn get_mut(&mut self, key: &Self::RefKey) -> Option<&mut Self::Value>;
 
-    fn remove(&mut self, key: &Self::RefKey) -> Self::Value;
+    fn remove(&mut self, key: &Self::RefKey) -> Option<Self::Value>;
 }
 
 pub trait Import {